feature: add diskquota support regular expression
Add diskquota support regular expression.

We can use pouch run --disk-quota="key=value" to set quota for
container. This pr can make it support to use regular expression key,
such as: --disk-quota=".=1g", it means all the quota of mount points
are'1g', the priority of key'.'is the lowest. If it has no'key', just
have'value', such as: --disk-quota=1g, it means the quota of rootfs is
1g. If the mountpoint is a volume, the quota of volume is base on
volume's size, the disk-quota is not take effect on a volume.

Signed-off-by: Rudy Zhang <rudyflyzhang@gmail.com>