bugfix: align with docker v1.24 API
The docker exec api uses custom stream bytes if the tty is false. In order
to align with docker API, we need to use
github.com/docker/docker/pkg/stdcopy package to modify stream bytes.

In order to make the cri CI happy, we should skip the related `echo`
test.

Signed-off-by: Wei Fu <fhfuwei@163.com>