refactor: image daemon mgr
Redesign the imageStore in image mgr and make it stable.

In the imageStore, we design primary reference and searchable reference.
The primary reference is used for pull in containerd and searchable
references are the alias the primary reference.

We can use different references to find the same image so that it
provides the better UEX command line.

Signed-off-by: Wei Fu <fhfuwei@163.com>