bugfix: make ringbuffer right
As the original design says, the ringbuffer uses two pointer to represent
push and pop position respectively in the circle list. The ringbuffer
allows to override oldest data if the buffer is full.

However, the Close action will be hang if the push pointer doesn't equal to
pop pointer. Please check the case.

```
E = empty, H = has data

At this time, the circle has full buffer, which means each node contains
the data.

   0H -> 1H -> 2H -> 3H -> 4H -> 0H
               ^     ^
               |     |
               |   push pointer
              pop pointer

In the following goroutine scheduled, the pop action will consume all
the data.

   0E -> 1E -> 2E -> 3E -> 4E -> 0E
               ^     ^
               |     |
               |   push pointer
              pop pointer

However, if there are not enough data to make the push pointer equals to
the pop pointer, the Close action will be hang.
```

In order to make the ringbuffer right, use FIFO queue to as buffer. The
front pointer contains the oldest data. When the buffer is full and
there is new data coming, the ringbuffer can drop the front one and push
new data into the back.

However, the ringbuffer still contains the data after closed. The
containerio should drain all the data and write it into the backend.

Signed-off-by: Wei Fu <fhfuwei@163.com>