enhance: add new formatter for syslog
A lot of log collectors consumes log data and indexs it by the timestamp
in millisecond format. In order to make the collector to consume data in
the order, we need to add the sequence number in the log. Simplify the
case, PouchContainer will uses the nano timestamp as sequene number.

Signed-off-by: Wei Fu <fhfuwei@163.com>