bugfix: fix can't set macaddress for endpoint
When user set the mac address in container's config:
`container.Config.MacAddress`, now it will not take effect.

The method is setting the mac address into the endpoint's options, it
will set into network device by libnetwork or it's remote plugins.

Add testcase: `TestCreateWithMacAddress` to check the mac address is set
into container's device or not.

Signed-off-by: Rudy Zhang <rudyflyzhang@gmail.com>