bugfix: cleanup containerd container before do restart policy
If the cleanup action is taken before restart policy, the restart action
will be failed because new container keeps the same name with the old
one. We should cleanup the old one before the restart policy.

In order to make it happen, we add the cleanup function in the hooks.

Signed-off-by: Wei Fu <fuweid89@gmail.com>