remove unused member of daemon config
ListenCRI which is the member of daemon config, is unused.

ListenCRI was first added in commit 05e492390. Commit 05e492390
was pushed on Jan 12, and merged on Mar 6.

ListenCRI was deleted in commit 05e492390, which was merged on
Jan 16.

So, it was a mistake that the deleted ListenCRI was added back
by the merge of 05e492390.

Signed-off-by: Wang Rui <baijia.wr@antfin.com>