copy: allow to mount /dev /tmp/ /proc
The /dev/ /tmp /proc are default mount point from oci spec. They are not
in the container mount points. When we pass the /dev/terminal-log as
dest, it will mount after the /dev, which handed by runC in the special
namespace. But when we try to mount for copy, we don't have /dev folder
or sub folder, we need to create parent folder for the target or empty
file.

Signed-off-by: Wei Fu <fuweid89@gmail.com>