bugfix: removing image by ID should conflict with running container.
If image is used by a running container and this image has another tag,
there are two reference of this image. Remove the image by ID without
"--force" will success that is unexpected.

This patch will fix it. We should check if any containers using this
image ID. And add tests.

Signed-off-by: Wang Rui <baijia.wr@antfin.com>