feat: support kill command
In fact, we consider carefully about add a kill command for pouch, and finally we think it's a good idea, because much time when we want to stop a container immediately, we can use kill command instead of stop command(which have default timeout 10s).
On the other hand, docker already support kill command!